refactor(structure): segregando serviços de inserção de imagens e padronizando criação de vehicle com imagens _V2
