+++ src/AdsetManagement.Application/DTOs/Requests/OtherOptionsRequest.cs
@@ -1,0 +1,9 @@
+namespace AdsetManagement.Application.DTOs.Requests;
+
+public class OtherOptionsRequest
+{
+    public bool? ArCondicionado { get; set; }
+    public bool? Alarme { get; set; }
+    public bool? Airbag { get; set; }
+    public bool? ABS { get; set; }
+}
